VueDjAgency/requirements.txt
asgiref==3.4.1
black==22.3.0
click==8.0.4
dataclasses==0.8
Django==3.2.13
djangorestframework==3.13.1
importlib-metadata==4.8.3
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
pkg_resources==0.0.0
platformdirs==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==1.2.3
typed-ast==1.5.4
typing_extensions==4.1.1
zipp==3.6.0
